requirements.txt
django>=1.7
python-social-auth
django-debug-toolbar
